Add files via upload
As per Wikipedia, "Supervised learning is the machine learning task of learning a function that maps an input to an output based on example input-output pairs. It infers a function from labeled training data consisting of a set of training examples. "

Here, we explore the Linear Regression Algorithm to predict the score of a student when the hours he spent on studying is 9.25 hours.